refactor: @Authentication으로 인증객체 주입방식 수정 & 디렉토리 구조 개선(#13)
