+++ src/main/java/gravit/code/common/auth/oauth/dto/NaverUserInfo.java
@@ -1,4 +1,4 @@
-package gravit.code.auth.oauth.dto;
+package gravit.code.common.auth.oauth.dto;
 
 import java.util.Map;
