+++ src/main/java/gravit/code/common/auth/oauth/dto/OAuthUserInfo.java
@@ -1,4 +1,4 @@
-package gravit.code.auth.oauth.dto;
+package gravit.code.common.auth.oauth.dto;
 
 public interface OAuthUserInfo {
