+++ src/main/java/gravit/code/common/auth/oauth/startegy/GoogleOAuthResponseStrategy.java
@@ -1,8 +1,8 @@
-package gravit.code.auth.oauth.startegy;
+package gravit.code.common.auth.oauth.startegy;
 
 
-import gravit.code.auth.oauth.dto.GoogleUserInfo;
-import gravit.code.auth.oauth.dto.OAuthUserInfo;
+import gravit.code.common.auth.oauth.dto.GoogleUserInfo;
+import gravit.code.common.auth.oauth.dto.OAuthUserInfo;
 import org.springframework.stereotype.Component;
 
 import java.util.Map;
