+++ src/main/java/gravit/code/common/auth/oauth/startegy/NaverOAuthResponseStrategy.java
@@ -1,8 +1,8 @@
-package gravit.code.auth.oauth.startegy;
+package gravit.code.common.auth.oauth.startegy;
 
 
-import gravit.code.auth.oauth.dto.NaverUserInfo;
-import gravit.code.auth.oauth.dto.OAuthUserInfo;
+import gravit.code.common.auth.oauth.dto.NaverUserInfo;
+import gravit.code.common.auth.oauth.dto.OAuthUserInfo;
 import org.springframework.stereotype.Component;
 
 import java.util.Map;
