+++ src/main/java/gravit/code/common/auth/oauth/startegy/OAuthResponseFactory.java
@@ -1,6 +1,6 @@
-package gravit.code.auth.oauth.startegy;
+package gravit.code.common.auth.oauth.startegy;
 
-import gravit.code.auth.oauth.dto.OAuthUserInfo;
+import gravit.code.common.auth.oauth.dto.OAuthUserInfo;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
