+++ src/main/java/gravit/code/common/auth/oauth/startegy/OAuthResponseStrategy.java
@@ -1,6 +1,6 @@
-package gravit.code.auth.oauth.startegy;
+package gravit.code.common.auth.oauth.startegy;
 
-import gravit.code.auth.oauth.dto.OAuthUserInfo;
+import gravit.code.common.auth.oauth.dto.OAuthUserInfo;
 
 import java.util.Map;
