+++ src/main/java/gravit/code/common/config/WebClientConfig.java
@@ -1,4 +1,4 @@
-package gravit.code.auth.config;
+package gravit.code.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
