+++ src/main/java/gravit/code/learning/controller/LearningController.java
@@ -1,6 +1,7 @@
 package gravit.code.learning.controller;
 
 import gravit.code.chapterProgress.dto.response.ChapterInfoResponse;
+import gravit.code.common.auth.oauth.LoginUser;
 import gravit.code.learning.dto.request.LearningResultSaveRequest;
 import gravit.code.learning.facade.LearningFacade;
 import gravit.code.unitProgress.dto.response.UnitPageResponse;
@@ -9,6 +10,7 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -20,22 +22,20 @@ public class LearningController {
 
     private final LearningFacade learningFacade;
 
-    // TODO userId -> 머지 후, @AuthenticationPrincipal로 전환
-
-    @GetMapping("/chapters/{userId}")
-    public ResponseEntity<List<ChapterInfoResponse>> getAllChapters(@PathVariable("userId") Long userId){
-        return ResponseEntity.status(HttpStatus.OK).body(learningFacade.getAllChapters(userId));
+    @GetMapping("/chapters")
+    public ResponseEntity<List<ChapterInfoResponse>> getAllChapters(@AuthenticationPrincipal LoginUser loginUser) {
+        return ResponseEntity.status(HttpStatus.OK).body(learningFacade.getAllChapters(loginUser.getId()));
     }
 
-    @GetMapping("/units/{chapterId}/{userId}")
-    public ResponseEntity<List<UnitPageResponse>> getAllUnits(@PathVariable("userId") Long userId,
+    @GetMapping("/{chapterId}/units")
+    public ResponseEntity<List<UnitPageResponse>> getAllUnits(@AuthenticationPrincipal LoginUser loginUser,
                                                               @PathVariable("chapterId") Long chapterId){
-        return ResponseEntity.status(HttpStatus.OK).body(learningFacade.getAllUnits(userId, chapterId));
+        return ResponseEntity.status(HttpStatus.OK).body(learningFacade.getAllUnits(loginUser.getId(), chapterId));
     }
 
-    @PostMapping("/results/{userId}")
-    public ResponseEntity<UserLevelResponse> saveLearningResult(@PathVariable Long userId,
+    @PostMapping("/results")
+    public ResponseEntity<UserLevelResponse> saveLearningResult(@AuthenticationPrincipal LoginUser loginUser,
                                                                 @Valid@RequestBody LearningResultSaveRequest request){
-        return ResponseEntity.status(HttpStatus.OK).body(learningFacade.saveLearningProgress(userId, request));
+        return ResponseEntity.status(HttpStatus.OK).body(learningFacade.saveLearningProgress(loginUser.getId(), request));
     }
 }
